Move Dbxfs to a special folder
Now it seems like a module

Version up to 0.4.1
