+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "codewarsbot",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "description": "Bot for tracking katas at Codewars in Telegram",
   "main": "index.js",
   "scripts": {
